requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.11.4
transformers==4.51.3
torch==2.6.0
fugashi[unidic-lite]==1.3.0
jaconv==0.3.4
mysql-connector-python==8.2.0
requests==2.32.3
regex==2024.11.6
pandas==2.2.3
tqdm==4.67.1
python-multipart==0.0.6
sentencepiece
peft
